@@ -1,53 +1,64 @@
 {
 	"name": "developer-hub-crm",
-	"version": "0.0.1",
 	"private": true,
+	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"test": "npm run test:integration && npm run test:unit",
+		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write .",
-		"test:integration": "playwright test",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint .",
 		"test:unit": "vitest",
-		"drizzle-generate": "drizzle-kit generate:sqlite",
-		"migrate": "tsm ./src/lib/server/migrate.ts",
-		"studio": "pnpm drizzle-kit studio"
+		"test": "npm run test:unit -- --run && npm run test:e2e",
+		"test:e2e": "playwright test",
+		"db:push": "drizzle-kit push",
+		"db:migrate": "drizzle-kit migrate",
+		"db:studio": "drizzle-kit studio"
 	},
 	"devDependencies": {
-		"@libsql/client": "^0.3.2",
-		"@lucia-auth/adapter-sqlite": "^2.0.0",
-		"@playwright/test": "^1.37.1",
-		"@sveltejs/adapter-auto": "^2.1.0",
-		"@sveltejs/kit": "^1.24.0",
-		"@tailwindcss/typography": "^0.5.9",
-		"@typescript-eslint/eslint-plugin": "^6.5.0",
-		"@typescript-eslint/parser": "^6.5.0",
-		"autoprefixer": "^10.4.15",
-		"daisyui": "^3.6.4",
-		"dotenv": "^16.3.1",
-		"drizzle-kit": "^0.19.13",
-		"drizzle-orm": "^0.28.5",
-		"drizzle-zod": "^0.5.1",
-		"eslint": "^8.48.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.33.0",
-		"lucia": "^2.4.1",
-		"postcss": "^8.4.29",
-		"postcss-load-config": "^4.0.1",
-		"prettier": "^3.0.3",
-		"prettier-plugin-svelte": "^3.0.3",
-		"svelte": "^4.2.0",
-		"svelte-check": "^3.5.1",
-		"tailwindcss": "^3.3.3",
-		"tslib": "^2.6.2",
-		"tsm": "^2.3.0",
-		"typescript": "^5.2.2",
-		"vite": "^4.4.9",
-		"vitest": "^0.32.2"
+		"@eslint/compat": "^1.2.7",
+		"@eslint/js": "^9.23.0",
+		"@playwright/test": "^1.51.1",
+		"@sveltejs/adapter-node": "^5.2.12",
+		"@sveltejs/kit": "^2.20.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/typography": "^0.5.16",
+		"@tailwindcss/vite": "^4.0.15",
+		"@testing-library/jest-dom": "^6.6.3",
+		"@testing-library/svelte": "^5.2.7",
+		"daisyui": "^5.0.9",
+		"date-fns": "^4.1.0",
+		"drizzle-kit": "^0.30.5",
+		"eslint": "^9.23.0",
+		"eslint-config-prettier": "^10.1.1",
+		"eslint-plugin-svelte": "^3.3.3",
+		"globals": "^16.0.0",
+		"jsdom": "^26.0.0",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"svelte": "^5.25.2",
+		"svelte-check": "^4.1.5",
+		"tailwindcss": "^4.0.15",
+		"typescript": "^5.8.2",
+		"typescript-eslint": "^8.27.0",
+		"vite": "^6.2.2",
+		"vitest": "^3.0.9"
 	},
-	"type": "module"
+	"dependencies": {
+		"@libsql/client": "^0.15.0",
+		"@node-rs/argon2": "^2.0.2",
+		"@oslojs/crypto": "^1.0.1",
+		"@oslojs/encoding": "^1.1.0",
+		"drizzle-orm": "^0.41.0"
+	},
+	"pnpm": {
+		"onlyBuiltDependencies": [
+			"esbuild"
+		]
+	}
 }
