@@ -1,13 +1,16 @@
-import { Config } from 'drizzle-kit'
-import * as dotenv from 'dotenv'
-dotenv.config()
+import { defineConfig } from 'drizzle-kit';
+if (!process.env.DATABASE_URL)
+	throw new Error('DATABASE_URL is not set');
 
-export default {
-	schema: './src/lib/server/schema',
-	out: './migrations',
-	driver: 'turso',
+export default defineConfig({
+	schema: './src/lib/server/db/schema.ts',
+
 	dbCredentials: {
-		url: process.env.DATABASE_URL || '',
-		authToken: process.env.DATABASE_AUTH_TOKEN || '',
+		url: process.env.DATABASE_URL,
+		authToken: process.env.DATABASE_AUTH_TOKEN,
 	},
-} satisfies Config
+
+	verbose: true,
+	strict: true,
+	dialect: 'turso',
+});
