@@ -1,7 +1,30 @@
-import { auth } from '$lib/server/lucia'
-import type { Handle } from '@sveltejs/kit'
+import type { Handle } from '@sveltejs/kit';
+import * as auth from '$lib/server/auth.js';
 
-export const handle: Handle = async ({ event, resolve }) => {
-	event.locals.auth = auth.handleRequest(event)
-	return await resolve(event)
-}
+const handleAuth: Handle = async ({ event, resolve }) => {
+	const sessionToken = event.cookies.get(auth.sessionCookieName);
+	if (!sessionToken) {
+		event.locals.user = null;
+		event.locals.session = null;
+		return resolve(event);
+	}
+
+	const { session, user } =
+		await auth.validateSessionToken(sessionToken);
+	if (session) {
+		auth.setSessionTokenCookie(
+			event,
+			sessionToken,
+			session.expiresAt,
+		);
+	} else {
+		auth.deleteSessionTokenCookie(event);
+	}
+
+	event.locals.user = user;
+	event.locals.session = session;
+
+	return resolve(event);
+};
+
+export const handle: Handle = handleAuth;
